Refactored change notification into CBLChangeNotifier
CBLChangeNotifier manages a set of CBLChangeListenerTokens, and
broadcasts change objects to them.
Made CBLChangeListenerToken more self-contained by having it call the
block itself.
Replaced the listener-handling code in CBLDatabase, CBLReplicator and
CBLLiveQuery with CBLChangeNotifier.
Created CBLDocumentChangeNotifier to manage per-document change
notification, so it can keep track of the C4DocumentObserver instead
of CBLDatabase.
Removed some old incorrect logic about not posting document-changed
notifications if the database is in a transaction.